+++ extensions/diffutil/CHANGELOG.md
@@ -5,6 +5,7 @@
 * `lifecycle-runtime-ktx` in combination with coroutines for lifecycle aware launch/cancellation of coroutines.
 * `diffSwapList` async method with lifecycleScope coroutine job launch for heavy data computation.
 * `cancelDiffSwapJob` method to cancel ongoing coroutine job.
+* `loadingStateListener` for async loading state callback.
 
 ### Removed
 * SmartAdapterBuilderExt with extension methods